+++ changelog.md
@@ -7,9 +7,29 @@
 -   [t]—test suite improvement
 -   [d]—docs improvement
 
-## 2.3.4 (WIP)
+
+## 2.3.7 (February 22, 2022)
+
+-   [f] Fix tests for self-ref foreign keys.
+-   [t] Run tests of PRs.
+
+
+## 2.3.6 (February 22, 2022)
+
+-   [f] Fix generic models for Nim <= 1.6.4. The same fix has been merged in ``std/macros``, in the devel branch of the compiler but will not be backported to 1.6 (see [#132](https://github.com/moigagoo/norm/issues/132)).
+-   [f] Fix self-referencing optional foreign key, handled manually through ``fk`` pragmas (see [#137](https://github.com/moigagoo/norm/issues/137) ).
+
+
+## 2.3.5 (January 17, 2022)
+
+-   [+] Add `sum` proc to calculate column value sum.
+
+
+## 2.3.4 (January 16, 2022)
 
 -   [+] Add the ability to define read-only models (see [#125](https://github.com/moigagoo/norm/issues/125)).
+-   [+] Add `exists` proc to check if a row exists (see [#115](https://github.com/moigagoo/norm/issues/115)).
+-   [r] Remove `nimibook` from dependencies to speed up package builds.
 
 
 ## 2.3.3 (December 26, 2021)